test(extension) Split `Extension` into smaller test suites.
