+++ tests/units/Extension/Constants.php
@@ -1,0 +1,36 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace Wasm\Tests\Units\Extension;
+
+use ReflectionExtension;
+use Wasm\Tests\Suite;
+
+class Constants extends Suite
+{
+    public function getTestedClassName()
+    {
+        return 'StdClass';
+    }
+
+    public function getTestedClassNamespace()
+    {
+        return '\\';
+    }
+
+    public function test_reflection_constants()
+    {
+        $this
+            ->given($reflection = new ReflectionExtension('wasm'))
+            ->when($result = $reflection->getConstants())
+            ->then
+                ->array($result)
+                    ->isEqualTo([
+                        'WASM_TYPE_I32' => 0,
+                        'WASM_TYPE_I64' => 1,
+                        'WASM_TYPE_F32' => 2,
+                        'WASM_TYPE_F64' => 3,
+                    ]);
+    }
+}
